init-install: add support for install image to Intel VROC raid disk
Intel VROC raid device named as /dev/mdXX, now it only can boot with
initramfs method because it need mdadm to do initial work in it, so do
not remove initrd for such kind device.

It also use "p" as partition prefix, adjust it to aviod use a wrong device name
durning install.

Signed-off-by: Liwei Song <liwei.song@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>